Updated de-ch person dataset
Updated the person name dataset for the de-ch locale.
For the first name, the dataset of the swiss statistical office
was taken. The data set was split by male/female names.
In each group the ones which occoured less than 60 times where taken out.
This lead to datasets of about 4000 names for males and females.
Of this lists a random sample of 2500 names was taken.

The source data of the surnames is from the swiss post, again, the data was
filtered such that only names that are used more than 25 times are present.
Then again, a random sample of 2500 names was taken of this list.