Updated to use `random.choices()` for random strings.
Available since Python 3.6, `random.choices()` is significantly faster
when generating a random string:

    In [1]: import random, string

    In [2]: characters = string.ascii_letters + string.digits

    In [3]: %timeit "".join(random.choice(characters) for _ in range(128))
    29.3 µs ± 87.2 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

    In [4]: %timeit "".join(random.choices(characters, k=128))
    7.92 µs ± 186 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

Note that this is backward incompatible with respect to reproducibility
because the two methods use different algorithms.

See https://docs.python.org/3/library/random.html#random.choices

Despite this, because of the improvement in performance - one of the
stated goals of this package, this seems like a reasonable trade off.