Updated to use `random.choices()` in more places.
Follow up to 27f144f9cb9263f1d4384be77e1f9b3c837190e9.

See also #1158.

I'm interested to see if this also resolves the problems with flaky
tests when calling either of the `Internet.query_string()` and
`Internet.query_parameters()` methods.

The former calls the latter and I suspect that the issue is caused by
the first call to `self.text.words()` returning duplicates which results
in overwriting a key in the dictionary that is constructed. As such, the
length of parameters is less than that requested.

I'm not sure if, by using `random.choices()` instead of many calls to
`random.choice()` in `Text.words()`, we'll have less chance of duplicate
values being chosen, but it seems prudent from a performance perspective
anyway.